updating simulation_pytorch example; updating Ray dep in `pyproject.toml`
