+++ pyproject.toml
@@ -54,7 +54,7 @@ protobuf = "^3.12.1"
 importlib-metadata = { version = "^1.4.0", markers = "python_version < '3.8'" }
 dataclasses = { version = "==0.6", markers = "python_version < '3.7'" }
 # Optional dependencies
-ray = { extras = ["default"], version = "==1.6.0", optional = true }
+ray = { extras = ["default"], version = "==1.9.2", optional = true }
 
 [tool.poetry.extras]
 simulation = ["ray"]
